% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_history_preparation.R
\name{bh_split_age_single_group}
\alias{bh_split_age_single_group}
\title{Prepare birth history data for a single age group}
\usage{
bh_split_age_single_group(
  data,
  id_columns,
  years,
  age_group_mo_start,
  age_group_mo_end
)
}
\arguments{
\item{data}{(\link[data.table:data.table]{data.table}) Full birth histories}

\item{id_columns}{(\code{character(N)}) Fields in \code{data} that uniquely identify individual
records}

\item{years}{(\code{integer(N)}) Years to consider}

\item{age_group_mo_start}{(\code{integer(1)}) Months of age completed when entering the
age group.}

\item{age_group_mo_end}{(\code{integer(1)}) Months of age when exiting the age group - e.g.
(6, 12) correponds to an age group of 6-11 months of completed age.}
}
\value{
Data subset to individuals who entered the age group during the specified years
}
\description{
Split out a single age group from full birth history data.
}
