% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_microdata_table.R
\name{clean_microdata_using_codebook}
\alias{clean_microdata_using_codebook}
\title{Clean a microdata table based on a codebook}
\usage{
clean_microdata_using_codebook(microdata, codebook)
}
\arguments{
\item{microdata}{(\link[data.table:data.table]{data.table}) Table containing raw DHS
microdata}

\item{codebook}{(\link[data.table:data.table]{data.table}) Codebook containing at least the
following fields:
\itemize{
\item "original_var": Variable as it is named in the raw microdata
\item "rename": Standardized variable name that will be returned after cleaning
\item "missing": Column representing "unknown"/"missing"/"NA" responses for a question.
This is a text column decoded by \code{\link[=interpret_text_as_integers]{interpret_text_as_integers()}}.
\item "include": For binary indicators, this column represents all "yes" responses. Yes
responses will be coded as 1, all other non-NA responses will be coded as 0. This
is a text column decoded by \code{\link[=interpret_text_as_integers]{interpret_text_as_integers()}}.
}}
}
\value{
Cleaned data.table containing the cleaned fields listed in codebook$rename
}
\description{
Rename and re-format columns from DHS microdata based on a codebook
}
\details{
Only columns listed in the codebook are retained. This function includes some
special processing for specific columns for more complex variables.
}
