% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_history_preparation.R
\name{cmc_to_year}
\alias{cmc_to_year}
\title{Convert CMC to year}
\usage{
cmc_to_year(cmc)
}
\arguments{
\item{cmc}{Date in CMC}
}
\value{
Date in integer year
}
\description{
Convert CMC to year
}
