% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acled_data.R
\name{download_acled_data}
\alias{download_acled_data}
\title{Download ACLED data through the API}
\usage{
download_acled_data(
  country_name,
  api_key = load_environment_variable("ACLED_API_KEY"),
  api_email = load_environment_variable("ACLED_EMAIL"),
  years = 2015:2020
)
}
\arguments{
\item{country_name}{(\code{character(1)}) Standardized name of the country where ACLED data
will be downloaded}

\item{api_key}{(\code{character(1)}) ACLED API key. Default pulled from R environment
variable \code{ACLED_API_KEY}.}

\item{api_email}{(\code{character(1)}) ACLED API username (an email address). Default pulled
from R environment variable \code{ACLED_EMAIL}.}

\item{years}{(\code{integer(N)}, default 2015:2020) Years for which ACLED data will be
downloaded}
}
\value{
R \link{data.table} containing results for this country and year
}
\description{
Download geopositioned conflict data from ACLED using an API query
}
\details{
The api_key and api_email parameters can be set as environment variables, and
are called from environment variables by default. For more information, see
\link[=https://developer.acleddata.com/]{the ACLED developer dashboard}.
}
