% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dhs_microdata.R
\name{download_dhs_microdata}
\alias{download_dhs_microdata}
\title{Download DHS microdata}
\usage{
download_dhs_microdata(survey_id)
}
\arguments{
\item{survey_id}{(\code{character(1)}) The text ID of the survey from which data and GPS
will be downloaded. Can be found via StatCompiler.}
}
\value{
A list of \link{data.table}s containing survey datasets.
}
\description{
Download a subset of DHS microdata for a given country and year
}
\details{
This function requires credentials to access the DHS API, requested via
\link[=https://dhsprogram.com]{the DHS Program website} and loaded into the environment
with \link[rdhs:set_rdhs_config]{rdhs}.
}
