% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_power_set}
\alias{generate_power_set}
\title{Generate power set}
\usage{
generate_power_set(x)
}
\arguments{
\item{x}{Vector from which subsets will be generated}
}
\value{
List where each entry is a member of the power set
}
\description{
Generate the power set (all unique subsets) of a vector
}
\details{
Includes both the complete set and the empty set as options. Taken from this
\href{https://stackoverflow.com/a/18718066/22642017}{Stackoverflow solution}
}
