% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_loss_function}
\alias{get_loss_function}
\title{Get loss function}
\usage{
get_loss_function(name)
}
\arguments{
\item{name}{(\code{character(1)}) One of "L1", "L2", or "BCE"}
}
\value{
A function that takes two arguments (observed and predicted) and generates
total error for a model
}
\description{
Helper function to pull L1, L2, or binary cross-entropy loss functions
}
