% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dhs_microdata.R
\name{get_rdhs_cache_directory}
\alias{get_rdhs_cache_directory}
\title{Get the RDHS cache directory}
\usage{
get_rdhs_cache_directory()
}
\description{
Get the folder where RDHS datasets are cached
}
\details{
Will stop and request the user to set up their RDHS config if not already
initialized
}
