% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportanceTracker.R
\name{ImportanceTracker}
\alias{ImportanceTracker}
\title{Variable Importance Tracker}
\description{
Variable Importance Tracker

Variable Importance Tracker
}
\details{
Used to track feature importance across model iterations. Uses Welford's
algorithm to update mean and sum of squares
}
\seealso{
https://github.com/iancovert/sage/blob/master/sage/utils.py#L142

importance_std_dev()

importance_variance(), converged()

self$importance_std_dev() self$converged()

self$get_convergence_ratio()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n_var}}{(\code{integer(1)}) Number of variables}

\item{\code{importance}}{(\code{numeric(n_var)}) Running mean of estimated feature importance
for each variable}

\item{\code{sum_of_squares}}{(\code{numeric(n_var)}) Running sum of squares of estimated
feature importance for each variable}

\item{\code{n_obs}}{(\code{integer(N)}) Total number of model runs observed so far for each
variable}

\item{\code{convergence_threshold}}{(\code{numeric(1)}) threshold below which the importance
values have effectively converged.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportanceTracker-new}{\code{ImportanceTracker$new()}}
\item \href{#method-ImportanceTracker-update}{\code{ImportanceTracker$update()}}
\item \href{#method-ImportanceTracker-importance_variance}{\code{ImportanceTracker$importance_variance()}}
\item \href{#method-ImportanceTracker-importance_std_dev}{\code{ImportanceTracker$importance_std_dev()}}
\item \href{#method-ImportanceTracker-get_convergence_ratio}{\code{ImportanceTracker$get_convergence_ratio()}}
\item \href{#method-ImportanceTracker-converged}{\code{ImportanceTracker$converged()}}
\item \href{#method-ImportanceTracker-clone}{\code{ImportanceTracker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-new}{}}}
\subsection{Method \code{new()}}{
Create a new ImportanceTracker object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$new(n_var, convergence_threshold = 0.025)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_var}}{Number of variables for which importance will be tracked}

\item{\code{convergence_threshold}}{threshold below which the importance values have
effectively converged. Expressed as the ratio between the largest standard
deviation across the importance values (numerator) and the difference between the
highest and lowest importance values (denominator).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-update"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-update}{}}}
\subsection{Method \code{update()}}{
Update running mean and sum of squares using Welford's algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$update(scores, which_sampled = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scores}}{Absolute importance scores calculated from a single run}

\item{\code{which_sampled}}{(\code{integer(N)}, default NULL) Indices of which variables were
sampled for this run. If NULL, all variables were sampled}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-importance_variance"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-importance_variance}{}}}
\subsection{Method \code{importance_variance()}}{
Get variance for each importance observation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$importance_variance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric vector with length \code{n_var} giving variance of the importance
measure for each variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-importance_std_dev"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-importance_std_dev}{}}}
\subsection{Method \code{importance_std_dev()}}{
Get standard deviation for each importance observation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$importance_std_dev()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric vector with length \code{n_var} giving standard deviations of the
importance measure for each variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-get_convergence_ratio"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-get_convergence_ratio}{}}}
\subsection{Method \code{get_convergence_ratio()}}{
Get the 'convergence ratio' between the maximum standard deviation of
any observation and the difference between largest and smallest observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$get_convergence_ratio()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
numeric(1) for convergence ratio; compare to self$convergence_threshold
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-converged"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-converged}{}}}
\subsection{Method \code{converged()}}{
Determine whether the importance values have converged
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$converged()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
logical(1) value for whether the model has converged
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportanceTracker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportanceTracker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportanceTracker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
