% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_microdata_table.R
\name{interpret_text_as_integers}
\alias{interpret_text_as_integers}
\title{Interpret text as integers}
\usage{
interpret_text_as_integers(text)
}
\arguments{
\item{text}{(\code{character(1)}) Text string possibly encoding one or more numbers}
}
\value{
Converted integer vector
}
\description{
Convert one-line character strings into numeric vectors
}
\details{
Used to parse semicolon-separated strings used in the codebook
}
\seealso{
\link{clean_microdata_using_codebook}
}
