% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_environment_variable}
\alias{load_environment_variable}
\title{Load environment variable}
\usage{
load_environment_variable(var, fail_on_empty = TRUE)
}
\arguments{
\item{var}{(character) environment variable to load}

\item{fail_on_empty}{(boolean, default \code{TRUE}) throw an error message if the variable
is not defined?}
}
\value{
Environment variable, as a character object
}
\description{
wrapper for \code{Sys.getenv()} that fails on empty values
}
