% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalImputer.R
\name{MarginalImputer}
\alias{MarginalImputer}
\title{R6 class used for imputation in Shapley decomposition}
\description{
R6 class used for imputation in Shapley decomposition

R6 class used for imputation in Shapley decomposition
}
\details{
Stores the full model and fitting dataset. Predicts on subsets of the input
data, replacing removed features with their marginal distributions.
}
\seealso{
https://github.com/iancovert/sage/blob/master/sage/imputers.py#L44
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{(\code{predict}) Fitted model that can be used for prediction}

\item{\code{features_table}}{Table of all possible features to use for prediction}

\item{\code{features}}{(\code{character(N)}) Names of model features used in the full model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarginalImputer-new}{\code{MarginalImputer$new()}}
\item \href{#method-MarginalImputer-predict}{\code{MarginalImputer$predict()}}
\item \href{#method-MarginalImputer-clone}{\code{MarginalImputer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-new}{}}}
\subsection{Method \code{new()}}{
Create a new MarginalImputer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$new(model, features_table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Fitted statistical model from which new predictions can be made using
the \code{predict()} function.}

\item{\code{features_table}}{(\code{data.frame}) Table with columns corresponding to features
in the \code{model}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-predict}{}}}
\subsection{Method \code{predict()}}{
Fit the predictive model with a subset of features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$predict(feature_subset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_subset}}{(\code{character(N)}) A subset of features to use for this model
version. All other features will be marginalized out.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
