% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_statcompiler_validation.R
\name{mbg_statcompiler_validation}
\alias{mbg_statcompiler_validation}
\title{MBG statcompiler validation}
\usage{
mbg_statcompiler_validation(
  viz_dir,
  input_data,
  adm1_mbg_summaries,
  country_name,
  sc_indicator,
  survey_year,
  sc_denominator = 1,
  label_function = scales::comma
)
}
\arguments{
\item{viz_dir}{(\code{character(1)}) Directory where all validation results and figures will
be saved.}

\item{input_data}{(\code{data.table}) Table containing at least the fields 'rate',
'samplesize', 'admin1_code', and 'admin1_name'}

\item{adm1_mbg_summaries}{(\code{data.table}) Table containing at least the fields
'ADM1_NAME', 'ADM1_CODE', 'mean', 'lower', and 'upper'}

\item{country_name}{(\code{character(1)}) Full name of country}

\item{sc_indicator}{(\code{character(1)}) Statcompiler indicator code}

\item{survey_year}{(\code{integer(1)}) Year corresponding to DHS survey}

\item{sc_denominator}{(\code{numeric(1)}, default 1) Denominator for statcompiler estimates.
Used to make statcompiler estimates compatible with MBG results}

\item{label_function}{(\code{function}) Function converting a numeric vector into a set of
labels.}
}
\description{
Compare MBG estimates against StatCompiler by admin1 unit
}
