% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_visualizations.R
\name{mbg_visualizations}
\alias{mbg_visualizations}
\title{MBG visualizations}
\usage{
mbg_visualizations(
  viz_dir,
  input_data,
  mean_raster,
  ui_raster,
  admin_data_for_plotting,
  adm1_boundaries,
  mesh = NULL,
  stacked_covariates = NULL,
  plot_stackers = TRUE,
  title_base = "",
  color_scheme = RColorBrewer::brewer.pal("Spectral", n = 9),
  indicator_name = "Indicator",
  label_function = scales::comma
)
}
\arguments{
\item{viz_dir}{(\code{character(1)}) Directory where all visualizations will be saved}

\item{input_data}{(\code{data.table}) Table with at least the columns "x", "y", "rate", and
"samplesize".}

\item{mean_raster}{(\code{spatRaster}) Mean MBG estimates by pixel}

\item{ui_raster}{(\code{spatRaster}) Raster showing width of the 95\% uncertainty interval
for each pixel}

\item{admin_data_for_plotting}{(\code{sf}) sf object with at least the rows 'mean', 'lower',
and 'upper'. Used for both visualization and admin2 boundaries.}

\item{adm1_boundaries}{(\code{sf}) sf object containing admin1 boundaries}

\item{mesh}{(\verb{INLA mesh}, default NULL) INLA mesh object.}

\item{stacked_covariates}{(list of \code{SpatRaster} objects) List containing stacked
covariates, if stacked covariates should be plotted.}

\item{title_base}{(\code{character(1)}) Template for all plot titles.}

\item{color_scheme}{(\code{character(N)}) vector of color codes or names for a gradient
color scheme.}

\item{indicator_name}{(\code{character(1)}) Name of the outcome being plotted, for use in
some legend titles.}

\item{label_function}{(\code{function}) Function converting a numeric vector into a set of
labels.}
}
\description{
Run standard visualizations for MBG input data and results
}
