% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dhs_microdata.R
\name{pull_dhs_survey_metadata}
\alias{pull_dhs_survey_metadata}
\title{Get DHS survey metadata
Get DHS survey metadata}
\usage{
pull_dhs_survey_metadata(country, years = "most_recent", survey_types = "all")
}
\arguments{
\item{country}{(\code{character(1)}) Country name to match to DHS survey list.
Case-insensitive}

\item{years}{(\code{integer(N)} or 'most_recent') Survey years to match to DHS survey list.
If 'most_recent', the default, pulls the latest available DHS standard survey}

\item{survey_types}{(\code{character(N)}, default 'all') Survey types to match to the survey
list. If 'all', the default, does not subset by survey type}
}
\value{
Any DHS survey IDs matching the listed countries and years.
}
\description{
Pull DHS internal survey metadata based on country, year, and survey type
}
\details{
Wrapper for \link[rdhs:dhs_surveys]{rdhs}
}
