% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg_visualizations.R
\name{raster_to_table}
\alias{raster_to_table}
\title{Raster to Table}
\usage{
raster_to_table(raster_object)
}
\arguments{
\item{raster_object}{terra SpatRaster object}
}
\value{
Data.frame object
}
\description{
Helper function to convert a terra SpatRaster into a data.frame
}
