% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_geospatial_covariates.R
\name{reshape_dhs_covariates}
\alias{reshape_dhs_covariates}
\title{Prepare geospatial covariates for analysis}
\usage{
reshape_dhs_covariates(
  covs_table,
  measure_years,
  id_col = "DHSCLUST",
  drop_cols = c("DHSID", "GPS_Dataset", "DHSCC", "DHSYEAR", "SurveyID"),
  keep_covariates = NULL
)
}
\arguments{
\item{covs_table}{Table of geospatial covariates returned from DHS API}

\item{measure_years}{All years needed for this analysis.}

\item{id_col}{Cluster ID as listed in the covariates table}

\item{drop_cols}{Columns that are not either covariates or the cluster ID}

\item{keep_covariates}{(\code{character(N)}, default NULL) Which covariates should be kept
in the final dataset? If NULL, the default, keep all available covariates}
}
\value{
Data.table reshaped long by year, with missing years filled
}
\description{
Reshape geospatial covariates wide and fill missing years
}
