% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shapley_decomposition.R
\name{run_all_covariate_combinations}
\alias{run_all_covariate_combinations}
\title{Run all covariate combinations}
\usage{
run_all_covariate_combinations(
  model_data,
  outcome,
  predictors,
  model_type,
  loss_function,
  model_options = list()
)
}
\arguments{
\item{model_data}{Model dataset, cleaned of NA values and including both the outcome
and predictor variables}

\item{outcome}{(\code{character(1)}) Character naming the field of the outcome to be
predicted}

\item{predictors}{(\code{character(N)}) A vector of all candidate models to consider}

\item{model_type}{(\code{character(1)}) Machine learning method to use, passed to caret.}

\item{loss_function}{A function that generates average model error given two vectors,
the two observations and the predicted risks.}

\item{model_options}{(\code{list(N)}, defaults to an empty list) Named list containing the
names of tuning parameters to be passed to caret. If arguments are not passed, the
caret defaults will be used.}
}
\value{
A named list, with names formatted as "cov_a+cov_b+cov_c+..." giving the model
total error for each set of covariates referred to in the name
}
\description{
Run all covariate combinations
}
\details{
Run a predictive model using all possible combinations from a list of
covariates and return the loss function for each combination.
}
\seealso{
\link{get_loss_function} For defining the loss function, \link{vector_to_names} for
converting a vector of covariates to a character string
}
