% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shapley_decomposition.R
\name{run_shapley_decomposition}
\alias{run_shapley_decomposition}
\title{Run Shapley decomposition on a set of model results}
\usage{
run_shapley_decomposition(predictors, model_performance_list)
}
\arguments{
\item{predictors}{(\code{character(N)}) All candidate predictors used in a model}

\item{model_performance_list}{(\code{list(N)}) List of model performance (usually decrease
in model error) for all combinations of those predictors}
}
\value{
Named list with two items:
\itemize{
\item 'absolute': Shapley value decomposition of each predictor to model performance
\item 'normalized': Same decomposition normalized to sum to one
}
}
\description{
Run Shapley decomposition on a set of model results
}
\details{
Given a vector of all candidate predictors and the performance of models
using all combinations of those predictors, assign relative importance of each
covariate to overall model performance
}
\seealso{
\link{vector_to_names} for converting a vector of covariates to a character string
}
