% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vector_to_names}
\alias{vector_to_names}
\title{Vector to names}
\usage{
vector_to_names(x)
}
\arguments{
\item{x}{Arbitrary vector}
}
\value{
(\code{character(1)}) A uniquely-identifying string representing this vector
}
\description{
Convert a vector into a unique name for e.g. a list
}
