% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_history_preparation.R
\name{year_to_cmc}
\alias{year_to_cmc}
\title{Convert year to CMC}
\usage{
year_to_cmc(year)
}
\arguments{
\item{year}{Date in integer years}
}
\value{
CMC corresponding to January of that year
}
\description{
Convert year to CMC
}
