% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalImputer.R
\name{MarginalImputer}
\alias{MarginalImputer}
\title{R6 class used for imputation in Shapley decomposition}
\description{
R6 class used for imputation in Shapley decomposition

R6 class used for imputation in Shapley decomposition
}
\details{
Stores the full model and fitting dataset. Predicts on subsets of the input
data, replacing removed features with their marginal distributions.
}
\seealso{
https://github.com/iancovert/sage/blob/master/sage/imputers.py#L44

wrapper function for MarginalImputer$predict()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{(\code{predict}) Fitted model that can be used for prediction}

\item{\code{features_table}}{Table of all possible features to use for prediction}

\item{\code{features}}{(\code{character(N)}) Names of model features used in the full model that
can be toggled.}

\item{\code{default_features}}{(\code{character(N)}) Names of model features used in the full
model that cannot be toggled}

\item{\code{outcomes}}{(\code{numeric(nrow(features_table))}) Vector of outcomes corresponding
to each row of \code{features_table}}

\item{\code{loss_fun}}{(function) Loss function taking two numeric vectors of equal
length, outcomes and predicted outcomes, and returning the total model loss}

\item{\code{prediction_fun}}{(function) Function to create predictions based on two items,
the model object \code{object} and new data \code{newdata}. Default is the generic
\code{predict} function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarginalImputer-new}{\code{MarginalImputer$new()}}
\item \href{#method-MarginalImputer-predict}{\code{MarginalImputer$predict()}}
\item \href{#method-MarginalImputer-get_loss}{\code{MarginalImputer$get_loss()}}
\item \href{#method-MarginalImputer-clone}{\code{MarginalImputer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-new}{}}}
\subsection{Method \code{new()}}{
Create a new MarginalImputer object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$new(
  model,
  features_table,
  outcomes,
  loss_fun,
  prediction_fun = stats::predict,
  default_features = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Fitted statistical model from which new predictions can be made using
the \code{predict()} function.}

\item{\code{features_table}}{(\code{data.frame}) Table with columns corresponding to features
in the \code{model}}

\item{\code{outcomes}}{(\code{numeric(nrow(features_table))}) Vector of outcomes corresponding
to each row of \code{features_table}}

\item{\code{loss_fun}}{(function) Loss function taking two numeric vectors of equal
length, outcomes and predicted outcomes, and returning the total model loss}

\item{\code{prediction_fun}}{(function) Function to create predictions based on two items,
the model object \code{object} and new data \code{newdata}. Default is the generic
\code{predict} function.}

\item{\code{default_features}}{(\code{character(N)}, default \code{NULL}) The features that will
be included in every model permutation by default. Not listed in self$features
because they cannot be toggled}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-predict"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-predict}{}}}
\subsection{Method \code{predict()}}{
Fit the predictive model with a subset of features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$predict(feature_subset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_subset}}{(\code{character(N)}) A subset of features to use for this model
version. All other features will be marginalized out.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-get_loss"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-get_loss}{}}}
\subsection{Method \code{get_loss()}}{
Get total model loss for a subset of features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$get_loss(feature_subset, epsilon = 1e-06)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature_subset}}{(\code{character(N)}) A subset of features to use for this model
version. All other features will be marginalized out.}

\item{\code{epsilon}}{(\code{numeric(1)}) The amount that very small (or large) probability
estimates should be perturbed from zero (or one)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarginalImputer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarginalImputer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarginalImputer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
